back_end/requirements.txt
fastapi
uvicorn
schematics
pymongo
motor
pydantic[email]
passlib
bcrypt
python-multipart
python-jose